backend/requirements.txt
fastapi
uvicorn
pytest
cryptography
httpx
pypdf
google-generativeai
python-multipart
pillow
transformers>=4.40.0
torch
accelerate
sqlalchemy>=2.0
aiosqlite
pydantic[email]
python-dotenv
redis>=5.0.0
supabase>=2.0.0
PyJWT>=2.8.0
requests
Pillow
